Code cleanup and refactoring sample project for documentation purposes
