+++ example/src/main/java/com/globalaccelerex/androidposclientlibrary/ActivitySample.kt
@@ -10,23 +10,32 @@ import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.P
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.*
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity() {
+/**
+ * This activity shows an example of how each library function is used.
+ * Same principles apply when calling from a fragment
+ * */
+
+class ActivitySample : AppCompatActivity() {
 
     private val clientLib = GAClientLib.Builder()
-        .setCountryCode(Countries.KENYA)
+        .setCountryCode(Countries.NIGERIA)
         .build()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
+        //Make Key Exchange
+        key_exchange.setOnClickListener {
+            clientLib.makeKeyExchangeRequest(this)
+        }
+
+        //Make parameter request
         paramter_request.setOnClickListener {
             clientLib.makeParametersRequest(this)
         }
 
-        key_exchange.setOnClickListener {
-            clientLib.makeKeyExchangeRequest(this)
-        }
+        //Make Card Present Transactions (CP)
         transaction_request.setOnClickListener {
             clientLib.makeCardPresentTransactionRequest(
                 amount = 1.0,
@@ -35,14 +44,8 @@ class MainActivity : AppCompatActivity() {
                 callingComponent = this
             )
         }
-        mobile_money_request.setOnClickListener {
-            clientLib.makeMobileMoneyTransactionRequest(
-                phoneNumber = "08143051805",
-                amount = 1.0,
-                mobileOperator = MobileMoneyOperators.MTN,
-                callingComponent = this
-            )
-        }
+
+        //Make Card Not Present Transactions (CNP)
         cnp_purchase.setOnClickListener {
             clientLib.makeCardNotPresentTransactionRequest(
                 cardNumber = "5196010125746208",
@@ -53,30 +56,41 @@ class MainActivity : AppCompatActivity() {
                 customPrint = false
             )
         }
+
+        //Make Mobile Money Transactions (MM)
+        mobile_money_request.setOnClickListener {
+            clientLib.makeMobileMoneyTransactionRequest(
+                phoneNumber = "08143051805",
+                amount = 1.0,
+                mobileOperator = MobileMoneyOperators.MTN,
+                callingComponent = this
+            )
+        }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
 
+        //Note that the Request key will always match the transaction type with "REQUEST_CODE" appended to it.
+
         if (resultCode == Activity.RESULT_OK) {
-            if (requestCode == GaResponseKeys.PARAMETERS_REQUEST_CODE) {
-                val posInfo: PosInformation? = clientLib.getPosParametersResponse(data)
-                Log.e("PosInfo", "$posInfo")
-            }
-            if (requestCode == GaResponseKeys.CNP_PURCHASE_REQUEST_CODE) {
-                val posInfo: PosInformation? = clientLib.getPosParametersResponse(data)
-                Log.e("PosInfo", "$posInfo")
-            }
-            if (requestCode == GaResponseKeys.KEY_EXCHANGE_REQUEST_CODE) {
+            if (requestCode == GaRequestKeys.KEY_EXCHANGE_REQUEST_CODE) {
                 Log.e("Key exchange", "Key exchange successful!")
             }
-            if (requestCode == GaResponseKeys.CARD_PURCHASE_REQUEST_CODE) {
+            if (requestCode == GaRequestKeys.PARAMETERS_REQUEST_CODE) {
+                val posInfo: PosInformation? = clientLib.getPosParametersResponse(data)
+                Log.e("PosInfo", "$posInfo")
+            }
+            if (requestCode == GaRequestKeys.CP_PURCHASE_REQUEST_CODE) {
                 val cardResponseDetails = clientLib.getCardTransactionResponse(data)
                 Log.e("Purchase response", "$cardResponseDetails")
             }
-            if (requestCode == GaResponseKeys.MOBILE_MONEY_PURCHASE_REQUEST_CODE) {
-                val cardResponseDetails = clientLib.getCardTransactionResponse(data)
-                Log.e("Purchase response", "$cardResponseDetails")
+            if (requestCode == GaRequestKeys.CNP_PURCHASE_REQUEST_CODE) {
+                // CNP Response type hasn't been defined yet
+            }
+            if (requestCode == GaRequestKeys.MOBILE_MONEY_PURCHASE_REQUEST_CODE) {
+                val mobileMoneyTransactionResponse = clientLib.getMobileMoneyTransactionResponse(data)
+                Log.e("Purchase response", "$mobileMoneyTransactionResponse")
             }
         }
     }
