+++ example/src/main/java/com/globalaccelerex/androidposclientlibrary/PrintingSample.kt
@@ -1,4 +1,6 @@
 package com.globalaccelerex.androidposclientlibrary
 
-class WrongParameterTest {
+class PrintingSample {
+
+
 }
