+++ pos-client-library/src/main/java/com/globalaccelerex/globalaccelerexandroidposclientlibrary/baseAppUtils/KeyExchangeRequest.kt
@@ -6,7 +6,7 @@ import androidx.fragment.app.Fragment
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.TerminalInformation.GHANA_TERMINAL_MODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.TerminalInformation.NIGERIA_TERMINAL_MODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.TerminalInformation.TERMINAL_MODE
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaRequestKeys
 
 internal class KeyExchangeRequest {
 
@@ -15,10 +15,10 @@ internal class KeyExchangeRequest {
             NIGERIA_TERMINAL_MODE, GHANA_TERMINAL_MODE -> {
                 val intent = Intent(BaseAppConstants.KEY_EXCHANGE_REQUEST_INTENT_ADDRESS)
                 if (callingComponent is Activity) {
-                    callingComponent.startActivityForResult(intent, GaResponseKeys.KEY_EXCHANGE_REQUEST_CODE)
+                    callingComponent.startActivityForResult(intent, GaRequestKeys.KEY_EXCHANGE_REQUEST_CODE)
                 }
                 if (callingComponent is Fragment){
-                    callingComponent.startActivityForResult(intent, GaResponseKeys.KEY_EXCHANGE_REQUEST_CODE)
+                    callingComponent.startActivityForResult(intent, GaRequestKeys.KEY_EXCHANGE_REQUEST_CODE)
                 }
             }
         }
